backend/requirements.txt
# Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
pymysql==1.1.0
asyncmy==0.2.9  # MySQL 异步驱动，用于 SQLAlchemy 2.0 异步操作

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 认证与安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 工具库
python-dateutil==2.8.2
pytz==2023.3

# OCR和图像处理
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0

# 中文处理
jieba==0.42.1

# Word 文档解析（内容审查用）
python-docx==1.1.0

# 日志
loguru==0.7.2

# AI 模型（通义千问）
dashscope>=1.25.10

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.0
